requirements.txt
streamlit
ffmpeg-python
faster-whisper
yt-dlp
psycopg2-binary
sentence-transformers
torch
openai
